procedimientos

+++ database/migrations/2022_11_01_012221_create_materiales_table.php
@@ -19,7 +19,7 @@ class CreateMaterialesTable extends Migration
             $table->integer('Costo');
             $table->integer('Unidades');
             $table->integer('Uso_de_unidad');
-            $table->integer('Costo_por_uso');
+            $table->float('Costo_por_uso');
             $table->timestamps();
         });
     }
